feat(admin): pick filter values inside the filters menu
